simplrefq-cloud-optimized.py
import os
import logging
import asyncio
from typing import Dict, Any

import pytz
import requests
import pymongo
from dotenv import load_dotenv
from flask import Flask, request, jsonify
from aiohttp import ClientSession
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import (
    ApplicationBuilder, 
    CommandHandler, 
    CallbackQueryHandler, 
    ContextTypes
)
import firebase_admin
from firebase_admin import credentials, messaging
from apscheduler.schedulers.background import BackgroundScheduler
from pymongo import MongoClient

# Enhanced Logging Configuration
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.StreamHandler(),
        logging.FileHandler('app.log')
    ]
)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()

# Configuration Class for Centralized Management
class Config:
    TELEGRAM_BOT_TOKEN = os.getenv('TELEGRAM_BOT_TOKEN')
    MONGO_URI = os.getenv('MONGO_URI')
    DB_NAME = os.getenv('DB_NAME', 'Cluster0')
    WEBHOOK_URL = os.getenv('WEBHOOK_URL', 'https://your-domain.com/webhook')
    PORT = int(os.getenv('PORT', 8080))
    FIREBASE_CREDENTIALS_PATH = os.getenv('FIREBASE_CREDENTIALS_PATH')

# Enhanced MongoDB Connection with Robust Error Handling
class DatabaseManager:
    def __init__(self, uri: str, db_name: str):
        try:
            self.client = MongoClient(uri, serverSelectionTimeoutMS=5000)
            self.db = self.client[db_name]
            self.users_collection = self.db['users']
            self.tasks_collection = self.db['tasks']
            self.logs_collection = self.db['audit_logs']
            logger.info(f"Successfully connected to MongoDB database: {db_name}")
        except pymongo.errors.PyMongoError as e:
            logger.critical(f"Failed to connect to MongoDB: {e}")
            raise

# Firebase Notification Service
class NotificationService:
    def __init__(self, credentials_path: str):
        try:
            cred = credentials.Certificate(credentials_path)
            firebase_admin.initialize_app(cred)
        except Exception as e:
            logger.error(f"Firebase initialization error: {e}")

    def send_push_notification(self, token: str, title: str, body: str):
        try:
            message = messaging.Message(
                notification=messaging.Notification(title=title, body=body),
                token=token
            )
            response = messaging.send(message)
            logger.info(f"Push notification sent: {response}")
        except Exception as e:
            logger.error(f"Notification sending error: {e}")

# Enhanced Wallet Connect Integration
class WalletConnectService:
    def __init__(self, db_manager: DatabaseManager):
        self.db = db_manager

    async def connect_wallet(self, user_id: int, wallet_address: str):
        try:
            # Validate wallet address (basic check)
            if not self._validate_wallet_address(wallet_address):
                return {"error": "Invalid wallet address"}

            # Update user record with wallet
            result = self.db.users_collection.update_one(
                {"user_id": user_id},
                {"$set": {
                    "wallet_address": wallet_address,
                    "wallet_verified": True
                }}
            )

            if result.modified_count:
                return {"status": "Wallet connected successfully"}
            return {"error": "Could not update wallet"}

        except Exception as e:
            logger.error(f"Wallet connection error: {e}")
            return {"error": "Wallet connection failed"}

    def _validate_wallet_address(self, address: str) -> bool:
        # Basic wallet address validation
        return (address.startswith('0x') and 
                len(address) == 42 and 
                all(c in '0123456789ABCDEFabcdef' for c in address[2:]))

# Flask Application Setup
def create_flask_app(
    config: Config, 
    db_manager: DatabaseManager, 
    notification_service: NotificationService,
    wallet_connect_service: WalletConnectService
):
    app = Flask(__name__)

    @app.route('/webhook', methods=['POST'])
    async def webhook():
        # Telegram webhook handling
        data = request.json
        # Process webhook data...
        return jsonify({"status": "success"}), 200

    @app.route('/wallet/connect', methods=['POST'])
    async def connect_wallet():
        data = request.json
        user_id = data.get('user_id')
        wallet_address = data.get('wallet_address')
        
        result = await wallet_connect_service.connect_wallet(user_id, wallet_address)
        return jsonify(result), 200 if 'status' in result else 400

    # Additional routes...
    return app

# Telegram Bot Setup
async def setup_telegram_bot(
    config: Config, 
    db_manager: DatabaseManager, 
    notification_service: NotificationService
):
    application = ApplicationBuilder().token(config.TELEGRAM_BOT_TOKEN).build()
    
    # Add handlers
    application.add_handler(CommandHandler('start', start))
    application.add_handler(CallbackQueryHandler(handle_callback))
    
    # Configure webhook
    await application.bot.set_webhook(url=config.WEBHOOK_URL)
    
    return application

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Existing start logic
    await update.message.reply_text("Welcome to ReblTasks!")

async def handle_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Callback query handling
    query = update.callback_query
    await query.answer()
    # Implement callback logic

# Main Application Orchestrator
async def main():
    config = Config()
    
    # Initialize services
    db_manager = DatabaseManager(config.MONGO_URI, config.DB_NAME)
    notification_service = NotificationService(config.FIREBASE_CREDENTIALS_PATH)
    wallet_connect_service = WalletConnectService(db_manager)
    
    # Create Flask app
    flask_app = create_flask_app(
        config, 
        db_manager, 
        notification_service, 
        wallet_connect_service
    )
    
    # Setup Telegram bot
    telegram_bot = await setup_telegram_bot(
        config, 
        db_manager, 
        notification_service
    )
    
    # Run application
    logger.info("Application started successfully")

if __name__ == '__main__':
    asyncio.run(main())



